Move Analysis::AST -> Analysis, keep AST as alias
